Big update was added S3 and minio, refactoring and adding dependencies and abstract class

+++ products/model.py
@@ -3,6 +3,7 @@ from sqlalchemy import func
 from sqlalchemy.orm import mapped_column
 
 from datetime import datetime
+from decimal import Decimal
 
 from db import Base
 
@@ -12,7 +13,7 @@ class CreateProductOrm(Base):
 
     id:Mapped[int] = mapped_column(primary_key=True)
     name:Mapped[str]
-    price:Mapped[float]
+    price:Mapped[Decimal]
     quantity:Mapped[int]
     description:Mapped[str]
     image_url:Mapped[str] = mapped_column(nullable=False)
